requirements.txt
black>=25.1.0
celery>=5.5.3
dj-rest-auth>=4.0
Django>=5.2.3
django-allauth>=0.61.1
django-cors-headers>=4.7.0
djangorestframework>=3.15
drf-yasg>=1.21.10
mysqlclient>=2.2.4
pre-commit>=4.2.0
psycopg2-binary>=2.9.10
pytest>=8.4.1
pytest-cov>=6.2.1
pytest-django>=4.11.1
python-decouple>=3.8
python-dotenv>=1.0.1
redis>=6.2.0
